+++ ClashDiscord/database/RazBotDB_player.py
@@ -1,4 +1,4 @@
-import RazBotDB_Presets as preset
+import database.RazBotDB_Presets as preset
 
 
 class Player(object):
